Reorganize validator dir structure
Previously, all files were placed in validator dir directly to quickly
resolve a require() path resolution error.
